Aggregation
If the table has odd number of rows,for example 3 rows 1 2 3 then for 2nd row the number of rows below it is 1 and no of rows above is 1 . so the diff is 0 and the median is the 2 nd row latitude value.
If the table has even no of rows ex: 4 rows 1 2 3 4 ,then for row 2 the number of rows below it is 1 and no of rows above it is 2,the diff is -1 and for the row 3 the number of rows below it 2 and number of rows above it is 1 ,so the difference is 1. The median is the avg of the 3rd and 4th row latitude values.